# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP Basic Authentication. Click \"Authorize\" button and enter username/password configured in the server.",
        "title": "MID Bootstrap Server API",
        "contact": {
            "name": "MAG Support",
            "url": "https://github.com/your-org/mid-bootstrap-server"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.4.3"
    },
    "basePath": "/",
    "paths": {
...
        "/api/alerts/{id}": {
            "get": {
                "description": "Handles individual alert operations: GET retrieves alert details, POST performs actions (acknowledge, resolve, delete-agent), DELETE removes the alert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Get, update, or delete an alert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action to perform: acknowledge, resolve, or delete-agent (POST only)",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "description": "Resolution details for resolve action",
                        "name": "resolution",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert details (GET) or action result (POST/DELETE)",
                        "schema": {
                            "$ref": "#/definitions/types.Alert"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Handles individual alert operations: GET retrieves alert details, POST performs actions (acknowledge, resolve, delete-agent), DELETE removes the alert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Get, update, or delete an alert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action to perform: acknowledge, resolve, or delete-agent (POST only)",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "description": "Resolution details for resolve action",
                        "name": "resolution",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert details (GET) or action result (POST/DELETE)",
                        "schema": {
                            "$ref": "#/definitions/types.Alert"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/alerts/{id}/{action}": {
            "post": {
                "description": "Handles individual alert operations: GET retrieves alert details, POST performs actions (acknowledge, resolve, delete-agent), DELETE removes the alert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Get, update, or delete an alert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action to perform: acknowledge, resolve, or delete-agent (POST only)",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "description": "Resolution details for resolve action",
                        "name": "resolution",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert details (GET) or action result (POST/DELETE)",
                        "schema": {
                            "$ref": "#/definitions/types.Alert"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/audit": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns audit log entries with optional filtering by hostname or event type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get audit log",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of entries to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by hostname",
                        "name": "hostname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event type (approval, denial, token_delivered, etc.)",
                        "name": "event_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of audit entries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.AuditEntry"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/manual-bootstrap": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Generates a bootstrap token for a machine without requiring a bootstrap request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Generate manual bootstrap token",
                "parameters": [
                    {
                        "description": "Manual bootstrap request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ManualBootstrapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated bootstrap token",
                        "schema": {
                            "$ref": "#/definitions/types.BootstrapToken"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/ws": {
            "get": {
                "description": "Establishes a WebSocket connection for receiving real-time events. Events include: new_request, request_approved, request_denied, token_generated, new_alert, alert_acknowledged, alert_resolved. Messages are JSON formatted with 'type' and 'data' fields.",
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket connection for real-time updates",
                "responses": {
                    "101": {
                        "description": "Switching Protocols - WebSocket connection established",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - WebSocket upgrade failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.ManualBootstrapRequest": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "types.Alert": {
            "type": "object",
            "properties": {
                "acknowledged": {
                    "description": "Acknowledgement",
                    "type": "boolean"
                },
                "acknowledged_at": {
                    "type": "string"
                },
                "acknowledged_by": {
                    "type": "string"
                },
                "agent_id": {
                    "description": "Agent information",
                    "type": "string"
                },
                "agent_version": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "description": "JSON for extra info",
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_seen_at": {
                    "description": "For stale agent alerts",
                    "type": "string"
                },
                "message": {
                    "description": "Alert details",
                    "type": "string"
                },
                "new_version": {
                    "type": "string"
                },
                "old_version": {
                    "description": "For version change alerts",
                    "type": "string"
                },
                "resolution": {
                    "description": "e.g., \"deleted_old_version\", \"agent_recovered\"",
                    "type": "string"
                },
                "resolved": {
                    "description": "Resolution",
                    "type": "boolean"
                },
                "resolved_at": {
                    "type": "string"
                },
                "resolved_by": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/types.AlertSeverity"
                },
                "stale_duration_min": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/types.AlertType"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}